+++ tests/framework/validators/IpValidatorTest.php
@@ -47,18 +47,8 @@ class IpValidatorTest extends TestCase
         new IpValidator(['ipv4' => false, 'ipv6' => false]);
     }
 
-    public static function provideRangesForSubstitution(): array
-    {
-        return [
-            ['10.0.0.1', ['10.0.0.1']],
-            [['192.168.0.32', 'fa::/32', 'any'], ['192.168.0.32', 'fa::/32', '0.0.0.0/0', '::/0']],
-            [['10.0.0.1', '!private'], ['10.0.0.1', '!10.0.0.0/8', '!172.16.0.0/12', '!192.168.0.0/16', '!fd00::/8']],
-            [['private', '!system'], ['10.0.0.0/8', '172.16.0.0/12', '192.168.0.0/16', 'fd00::/8', '!224.0.0.0/4', '!ff00::/8', '!169.254.0.0/16', '!fe80::/10', '!127.0.0.0/8', '!::1', '!192.0.2.0/24', '!198.51.100.0/24', '!203.0.113.0/24', '!2001:db8::/32']],
-        ];
-    }
-
     /**
-     * @dataProvider provideRangesForSubstitution
+     * @dataProvider \yiiunit\framework\validators\providers\IpValidatorProvider::rangesForSubstitution
      * @param array $range
      */
     public function testRangesSubstitution($range, array $expectedRange): void
@@ -80,13 +70,8 @@ class IpValidatorTest extends TestCase
         $this->assertFalse($validator->validate('babe::cafe'));
     }
 
-    public static function provideBadIps(): array
-    {
-        return [['not.an.ip'], [['what an array', '??']], [123456], [true], [false], ['bad:forSure']];
-    }
-
     /**
-     * @dataProvider provideBadIps
+     * @dataProvider \yiiunit\framework\validators\providers\IpValidatorProvider::badIps
      * @param mixed $badIp
      */
     public function testValidateValueNotAnIP($badIp): void
@@ -97,7 +82,7 @@ class IpValidatorTest extends TestCase
     }
 
     /**
-     * @dataProvider provideBadIps
+     * @dataProvider \yiiunit\framework\validators\providers\IpValidatorProvider::badIps
      * @param mixed $badIp
      */
     public function testValidateModelAttributeNotAnIP($badIp): void